add custom divs
