+++ app/pages/mint/[mint].tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import Loading from '../../components/Loading';
 import { TokenView } from '../../components/TokenView';
-import { KNOWN_MINTS } from '../../utils/constants';
+import { KNOWN_MINTS } from '../../utils/knownMints';
 import { MintWithMetadata, tryGetTokenMetaData } from '../../utils/metaplex';
 import { tryGetPubKey, tryParseTokenMint } from '../../utils/solana';
 
@@ -12,6 +12,7 @@ export default function Home() {
     const [mint, setMint] = useState<PublicKey>();
     const [isLoading, setIsLoading] = useState<boolean>(true);
     const [tokenData, setTokenData] = useState<MintWithMetadata>();
+    const [customDiv, setCustomDiv] = useState<JSX.Element>();
     const router = useRouter();
     const { connection } = useConnection();
 
@@ -27,6 +28,7 @@ export default function Home() {
             let knownMint = Object.keys(KNOWN_MINTS).find(address => address === searchValue);
             if (knownMint) { document.body.style.backgroundColor = KNOWN_MINTS[knownMint].background }
             else { document.body.style.backgroundColor = "black" }
+            if (knownMint && KNOWN_MINTS[knownMint].customDiv) {setCustomDiv(KNOWN_MINTS[knownMint].customDiv)}
             const mintPubKey = tryGetPubKey(searchValue);
             if (!mintPubKey) { setIsLoading(false); return };
             const info = await connection.getAccountInfo(mintPubKey);
@@ -43,7 +45,7 @@ export default function Home() {
     return (
         // TO DO ADD More margin above body. (hitting the navbar rn)
         <>
-            {tokenData ? <TokenView tokenData={tokenData} /> :
+            {tokenData ? <>{customDiv ?? ''} <TokenView tokenData={tokenData} /></> :
                 !isLoading ? <div className='mid-notice'>invalid mint</div> :
                     <div className="mid-notice"><Loading show={isLoading} text={'Loading . . . '} /></div>}
         </>
